% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrecipLZ.R
\name{getPrecipLZ}
\alias{getPrecipLZ}
\title{Get the observations}
\usage{
getPrecipLZ(method = NULL, path = NULL, filename = NULL,
  modelPrecipLZ = NULL)
}
\arguments{
\item{method}{method for processing observations, "none", "lz"}

\item{path}{Directory where to get the files}

\item{filename}{Filename (path included) of the file to be read}

\item{modelPrecipLZ}{list(nbLevelZone,Plr,hfelt,midmetp)}
}
\description{
Get and process, if needed, the observations
}
\examples{
\dontrun{
getPrecipLZ()
}
}
\keyword{precipLZ}

