% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz.R
\name{lz}
\alias{lz}
\title{Precipiation field in to LZ precipitation field}
\usage{
lz(precip, modelPrecipLZ)
}
\arguments{
\item{precip}{precipitation field}

\item{modelPrecipLZ}{list(nbLevelZone,Plr,hfelt,midmetp)}
}
\description{
Precipiation field in to LZ precipitation field
}
\examples{
\dontrun{
lz()
}
}
\keyword{precipLZ}
