% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{perturbation}
\alias{perturbation}
\title{Precipiation field in to LZ precipitation field}
\usage{
perturbation(precip, modelPert)
}
\arguments{
\item{precip}{precipitation field}

\item{modelPert}{list(ditr,param)}
}
\description{
Precipiation field in to LZ precipitation field
}
\examples{
\dontrun{
perturbation()
}
}
\keyword{precipLZ}

