% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateX.R
\name{stateX}
\alias{stateX}
\title{Precipitation for each elevation zone}
\usage{
stateX(v, modelPrecipLZ)
}
\arguments{
\item{v}{precipitation value (scalar)}

\item{modelPrecipLZ}{list(nbLevelZone,Plr,hfelt,midmetp)}
}
\description{
The function \code{stateX()} processes the precipitation for each elevation zone
}
\examples{
\dontrun{
stateX()
}
}
\keyword{precipLZ}

