% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrecipLZ.R
\name{getPrecipLZ}
\alias{getPrecipLZ}
\title{Get the observations}
\usage{
getPrecipLZ(method = NULL, path = NULL, filename = NULL, ...)
}
\arguments{
\item{method}{method for processing observations, "lz", "processedNVE"}

\item{path}{Directory where to get the files}

\item{filename}{Filename (path included) of the file to be read}

\item{...}{options for the different methods}
}
\description{
Get and process, if needed, the observations
}
\examples{
\dontrun{
getPrecipLZ()
}
}
\keyword{precipLZ}
